Update the interface and add main page as the start page to ask user to register in order to use regulato
